Move useSetupStore hook to EGEL feature directory
Relocated the useSetupStore Zustand hook from the auth feature to the EGEL feature directory for better organization. Updated imports in SetupPage to reflect the new path. No functional changes were made.
